+++ Code/Store3Mapi/ScribeMapi_Contact.cpp
@@ -42,7 +42,7 @@ LPMESSAGE GMapiContact::Handle()
 	{
 		ULONG Type = 0;
 		IUnknown *Item = NULL;
-		HRESULT e = Parent->Handle()->OpenEntry(	Entry.Length(),
+		HRESULT e = Parent->Handle()->OpenEntry(	(ULONG)Entry.Length(),
 													(LPENTRYID)&Entry[0],
 													NULL,
 													MAPI_BEST_ACCESS,
